ci: add clippy, cargo-audit, and sdist-install gates; resolve all RustSec advisories
New CI jobs in build.yml:
- clippy --workspace --all-targets -D warnings (existing warnings fixed:
  redundant f64 casts in instant.rs, boxed the large Sgp4Source::Tle
  variant; two tides.rs physical constants annotated as false positives)
- cargo audit (RustSec advisory scan)
- sdist-install guard: build the sdist, pip install from it, import —
  the check that would have caught the source-unbuildable 0.20.0 sdist

Dependency updates to clear all 7 outstanding advisories the audit gate
surfaced: quick-xml 0.38 -> 0.41 (two high-severity XML DoS advisories;
OMM XML is untrusted input), pyo3/numpy 0.28 -> 0.29 (one downcast->cast
rename), and compatible cargo update bumps (rustls-webpki, anyhow, rand).

Verified locally: fmt/clippy/audit clean, 254 Rust tests (incl. doc
tests), 114 Python tests, sdist guard green in a clean venv.

+++ src/time/instant.rs
@@ -320,7 +320,7 @@ impl Instant {
             TimeScale::UT1 => {
                 // This will be approximately correct for computing ut1
                 let eop = crate::earth_orientation_params::eop_from_mjd_utc_or_zero(mjd);
-                let dut1 = eop[0] as f64;
+                let dut1 = eop[0];
                 Self::from_mjd_with_scale(mjd - dut1 / 86_400.0, TimeScale::UTC)
             }
             TimeScale::GPS => {
@@ -415,7 +415,7 @@ impl Instant {
             TimeScale::UT1 => {
                 let mjd_utc = self.as_mjd_utc();
                 let eop = crate::earth_orientation_params::eop_from_mjd_utc_or_zero(mjd_utc);
-                let dut1 = eop[0] as f64;
+                let dut1 = eop[0];
                 mjd_utc + dut1 / 86_400.0
             }
             TimeScale::TAI => (self.raw - Self::MJD_EPOCH.raw) as f64 / 86_400_000_000.0,
